Seção 9: DTOs & Data Annotations

+++ Back/src/CursoAngular.API/Controllers/EventosController.cs
@@ -1,14 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-using CursoAngular.Persistence;
-using CursoAngular.Domain;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using CursoAngular.Persistence.Contexto;
 using CursoAngular.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using CursoAngular.Application.Dtos;
 
 namespace CursoAngular.API.Controllers
 {
@@ -29,8 +24,9 @@ namespace CursoAngular.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum Evento econtrado.");
+                if (eventos == null) return NoContent();
 
+                
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -46,7 +42,7 @@ namespace CursoAngular.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Nenhum Evento econtrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -62,7 +58,7 @@ namespace CursoAngular.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por tema não econtrados.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -73,12 +69,12 @@ namespace CursoAngular.API.Controllers
             }
         }
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar Evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -89,12 +85,12 @@ namespace CursoAngular.API.Controllers
             }
         }
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar Evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -109,9 +105,12 @@ namespace CursoAngular.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                        Ok("Deletado") : 
-                       BadRequest("Evento não deletado.");
+                       throw new Exception("Ocorreu um problema não específico ao tentar deletar o Evento");
             }
             catch (Exception ex)
             {
